requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
Django==4.2.4
django-crispy-forms==2.0
djangorestframework==3.14.0
djlint==1.32.1
EditorConfig==0.12.3
flake8==6.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.9
json5==0.9.14
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
